+++ apps/explorer/test/explorer/chain_test.exs
@@ -50,6 +50,51 @@ defmodule Explorer.ChainTest do
     end
   end
 
+  describe "address_to_logs/2" do
+    test "fetches logs" do
+      address = insert(:address)
+
+      transaction1 =
+        :transaction
+        |> insert(to_address: address)
+        |> with_block()
+
+      insert(:log, transaction: transaction1, index: 1, address: address)
+
+      transaction2 =
+        :transaction
+        |> insert(from_address: address)
+        |> with_block()
+
+      insert(:log, transaction: transaction2, index: 2, address: address)
+
+      assert Enum.count(Chain.address_to_logs(address)) == 2
+    end
+
+    test "paginates logs" do
+      address = insert(:address)
+
+      transaction =
+        :transaction
+        |> insert(to_address: address)
+        |> with_block()
+
+      log1 = insert(:log, transaction: transaction, index: 1, address: address)
+
+      2..51
+      |> Enum.map(fn index -> insert(:log, transaction: transaction, index: index, address: address) end)
+      |> Enum.map(& &1.index)
+
+      paging_options1 = %PagingOptions{page_size: 1}
+
+      [_log] = Chain.address_to_logs(address, paging_options: paging_options1)
+
+      paging_options2 = %PagingOptions{page_size: 60, key: {transaction.block_number, transaction.index, log1.index}}
+
+      assert Enum.count(Chain.address_to_logs(address, paging_options: paging_options2)) == 50
+    end
+  end
+
   describe "address_to_transactions_with_rewards/2" do
     test "without transactions" do
       address = insert(:address)
@@ -3901,6 +3946,61 @@ defmodule Explorer.ChainTest do
       end
 
       refute Chain.contract_address?(to_string(hash), 1, json_rpc_named_arguments)
+    end
+  end
+
+  describe "staking_pools/3" do
+    test "validators staking pools" do
+      inserted_validator = insert(:address_name, primary: true, metadata: %{is_active: true, is_validator: true})
+      insert(:address_name, primary: true, metadata: %{is_active: true, is_validator: false})
+
+      options = %PagingOptions{page_size: 20, page_number: 1}
+
+      assert [gotten_validator] = Chain.staking_pools(:validator, options)
+      assert inserted_validator.address_hash == gotten_validator.address_hash
+    end
+
+    test "active staking pools" do
+      inserted_validator = insert(:address_name, primary: true, metadata: %{is_active: true})
+      insert(:address_name, primary: true, metadata: %{is_active: false})
+
+      options = %PagingOptions{page_size: 20, page_number: 1}
+
+      assert [gotten_validator] = Chain.staking_pools(:active, options)
+      assert inserted_validator.address_hash == gotten_validator.address_hash
+    end
+
+    test "inactive staking pools" do
+      insert(:address_name, primary: true, metadata: %{is_active: true})
+      inserted_validator = insert(:address_name, primary: true, metadata: %{is_active: false})
+
+      options = %PagingOptions{page_size: 20, page_number: 1}
+
+      assert [gotten_validator] = Chain.staking_pools(:inactive, options)
+      assert inserted_validator.address_hash == gotten_validator.address_hash
+    end
+  end
+
+  describe "staking_pools_count/1" do
+    test "validators staking pools" do
+      insert(:address_name, primary: true, metadata: %{is_active: true, is_validator: true})
+      insert(:address_name, primary: true, metadata: %{is_active: true, is_validator: false})
+
+      assert Chain.staking_pools_count(:validator) == 1
+    end
+
+    test "active staking pools" do
+      insert(:address_name, primary: true, metadata: %{is_active: true})
+      insert(:address_name, primary: true, metadata: %{is_active: false})
+
+      assert Chain.staking_pools_count(:active) == 1
+    end
+
+    test "inactive staking pools" do
+      insert(:address_name, primary: true, metadata: %{is_active: true})
+      insert(:address_name, primary: true, metadata: %{is_active: false})
+
+      assert Chain.staking_pools_count(:inactive) == 1
     end
   end
 end