+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Current
 
 ### Features
+- [#1963](https://github.com/poanetwork/blockscout/pull/1963) - added rinkeby theme and rinkeby logo
 - [#1959](https://github.com/poanetwork/blockscout/pull/1959) - added goerli theme and goerli logo
 - [#1928](https://github.com/poanetwork/blockscout/pull/1928) - pagination styles were updated
 - [#1948](https://github.com/poanetwork/blockscout/pull/1948) - added ropsten theme and ropsten logo
@@ -21,6 +22,8 @@
 - [#1920](https://github.com/poanetwork/blockscout/pull/1920) - fix: remove source code fields from list endpoint
 - [#1876](https://github.com/poanetwork/blockscout/pull/1876) - async calculate a count of blocks
 - [#1941](https://github.com/poanetwork/blockscout/pull/1941) - feat: add on demand fetching and stale attr to rpc
+- [#1956](https://github.com/poanetwork/blockscout/pull/1956) - add logs tab to address
+- [#1933](https://github.com/poanetwork/blockscout/pull/1933) - add eth_BlockNumber json rpc method
 - [#1952](https://github.com/poanetwork/blockscout/pull/1952) - feat: exclude empty contracts by default
 
 ### Fixes
@@ -55,6 +58,7 @@
 - [#1892](https://github.com/poanetwork/blockscout/pull/1892) - Remove temporary worker modules
 - [#1958](https://github.com/poanetwork/blockscout/pull/1958) - Default value for release link env var
 - [#1964](https://github.com/poanetwork/blockscout/pull/1964) - ALLOWED_EVM_VERSIONS env var
+- [#1975](https://github.com/poanetwork/blockscout/pull/1975) - add log index to transaction view
 
 ## 1.3.10-beta